requirements.txt
streamlit==1.42.1
python-dotenv==1.0.1
crewai>=0.11.0
langchain>=0.1.0
langchain-google-genai>=0.0.3
google-generativeai>=0.3.0
chromadb==0.3.26
pysqlite3-binary==0.5.4
tiktoken==0.5.2
setuptools
pyyaml
nest-asyncio==1.5.8
sentence-transformers==2.2.2
faiss-cpu==1.7.4